refactor: kratos library
