+++ packages/shared/src/components/auth/LoginForm.tsx
@@ -3,12 +3,17 @@ import { useMutation, useQuery } from 'react-query';
 import AuthContext from '../../contexts/AuthContext';
 import {
   getNodeValue,
-  initializeLogin,
   LoginPasswordParameters,
-  validatePasswordLogin,
+  ValidateLoginParams,
 } from '../../lib/auth';
 import { formToJson } from '../../lib/form';
 import { disabledRefetch } from '../../lib/func';
+import {
+  AuthFlow,
+  AuthSession,
+  initializeKratosFlow,
+  submitKratosFlow,
+} from '../../lib/kratos';
 import { Button } from '../buttons/Button';
 import { ClickableText } from '../buttons/ClickableText';
 import { TextField } from '../fields/TextField';
@@ -28,14 +33,24 @@ function LoginForm({
 }: LoginFormProps): ReactElement {
   const { onUpdateSession } = useContext(AuthContext);
   const formRef = useRef<HTMLFormElement>();
-  const { data } = useQuery('login', initializeLogin, { ...disabledRefetch });
-  const { mutateAsync: onPasswordLogin } = useMutation(validatePasswordLogin, {
-    onSuccess: onUpdateSession,
-  });
+  const { data: login } = useQuery(
+    'login',
+    () => initializeKratosFlow(AuthFlow.Login),
+    { ...disabledRefetch },
+  );
+  const { mutateAsync: onPasswordLogin } = useMutation(
+    (params: ValidateLoginParams) => submitKratosFlow(params),
+    {
+      onSuccess: ({ data }) => {
+        const session = data as AuthSession;
+        onUpdateSession(session);
+      },
+    },
+  );
   const onLogin: typeof onSuccessfulLogin = async (e) => {
     e.preventDefault();
     const form = formToJson<LoginFormParams>(formRef.current);
-    const { nodes, action } = data.ui;
+    const { nodes, action } = login.ui;
     const csrfToken = getNodeValue('csrf_token', nodes);
     const params: LoginPasswordParameters = { ...form, csrf_token: csrfToken };
     onPasswordLogin({ action, params });
