+++ packages/shared/src/hooks/useRegistration.ts
@@ -1,19 +1,23 @@
 import { useMemo } from 'react';
 import { QueryKey, useMutation, useQuery } from 'react-query';
 import {
-  AuthSession,
   errorsToJson,
-  initializeRegistration,
   RegistrationError,
-  InitializationData,
   RegistrationParameters,
-  validateRegistration,
-  ValidateRegistrationParams,
   getNodeByKey,
   getNodeValue,
+  ValidateRegistrationParams,
 } from '../lib/auth';
 import { fallbackImages } from '../lib/config';
 import { disabledRefetch } from '../lib/func';
+import {
+  AuthFlow,
+  AuthSession,
+  InitializationData,
+  initializeKratosFlow,
+  submitKratosFlow,
+  SuccessfulRegistrationData,
+} from '../lib/kratos';
 
 type ParamKeys = keyof RegistrationParameters;
 
@@ -47,41 +51,45 @@ const useRegistration = ({
 }: UseRegistrationProps): UseRegistration => {
   const { data: registration, isLoading: isQueryLoading } = useQuery(
     key,
-    initializeRegistration,
+    () => initializeKratosFlow(AuthFlow.Registration),
     { ...disabledRefetch },
   );
   const {
     mutateAsync: validate,
     status,
     isLoading: isMutationLoading,
-  } = useMutation(validateRegistration, {
-    onSuccess: ({ data, error, redirect }, params) => {
-      if (data) {
-        return onValidRegistration?.(data.session, params);
-      }
+  } = useMutation(
+    (params: ValidateRegistrationParams) => submitKratosFlow(params),
+    {
+      onSuccess: ({ data, error, redirect }, params) => {
+        const successfulData = data as SuccessfulRegistrationData;
+        if (successfulData) {
+          return onValidRegistration?.(successfulData.session, params);
+        }
 
-      if (redirect) {
-        return onRedirect(redirect);
-      }
+        if (redirect) {
+          return onRedirect(redirect);
+        }
 
-      // probably csrf token issue and definitely not related to forms data
-      if (!error.ui) {
-        return null;
-      }
+        // probably csrf token issue and definitely not related to forms data
+        if (!error.ui) {
+          return null;
+        }
 
-      const emailExists = error?.ui?.messages?.find(
-        (message) => message.id === EMAIL_EXISTS_ERROR_ID,
-      );
-      if (emailExists) {
-        return onInvalidRegistration?.({
-          'traits.email': 'Email is already taken!',
-        });
-      }
+        const emailExists = error?.ui?.messages?.find(
+          (message) => message.id === EMAIL_EXISTS_ERROR_ID,
+        );
+        if (emailExists) {
+          return onInvalidRegistration?.({
+            'traits.email': 'Email is already taken!',
+          });
+        }
 
-      const json = errorsToJson<ParamKeys>(error);
-      return onInvalidRegistration?.(json);
+        const json = errorsToJson<ParamKeys>(error);
+        return onInvalidRegistration?.(json);
+      },
     },
-  });
+  );
 
   const onValidateRegistration = async (values: RegistrationParameters) => {
     const { nodes, action } = registration.ui;
